requirements.txt
# RefuTrack - Dependencias
Flask==3.1.2
Werkzeug==3.1.3
waitress==3.0.2
pywebview==6.0



